+++ tools/daemon/command.do.pl
@@ -136,10 +136,10 @@ sub getAnswerOk {
 # =================================================================================================
 
 if( !GetOptions(
-	"help!"				=> \$ep->{run}{help},
-	"colored!"			=> \$ep->{run}{colored},
-	"dummy!"			=> \$ep->{run}{dummy},
-	"verbose!"			=> \$ep->{run}{verbose},
+	"help!"				=> sub { $ep->runner()->help( @_ ); },
+	"colored!"			=> sub { $ep->runner()->colored( @_ ); },
+	"dummy!"			=> sub { $ep->runner()->dummy( @_ ); },
+	"verbose!"			=> sub { $ep->runner()->verbose( @_ ); },
 	"json=s"			=> \$opt_json,
 	"name=s"			=> \$opt_name,
 	"port=i"			=> sub {