+++ tools/daemon/vars.do.pl
@@ -62,10 +62,10 @@ sub listConfdir {
 # =================================================================================================
 
 if( !GetOptions(
-	"help!"				=> \$ep->{run}{help},
-	"colored!"			=> \$ep->{run}{colored},
-	"dummy!"			=> \$ep->{run}{dummy},
-	"verbose!"			=> \$ep->{run}{verbose},
+	"help!"				=> sub { $ep->runner()->help( @_ ); },
+	"colored!"			=> sub { $ep->runner()->colored( @_ ); },
+	"dummy!"			=> sub { $ep->runner()->dummy( @_ ); },
+	"verbose!"			=> sub { $ep->runner()->verbose( @_ ); },
 	"confDirs!"			=> \$opt_confDirs )){
 
 		msgOut( "try '".$ep->runner()->command()." ".$ep->runner()->verb()." --help' to get full usage syntax" );