+++ tools/ttp/vars.do.pl
@@ -147,10 +147,10 @@ sub listSitespec {
 # =================================================================================================
 
 if( !GetOptions(
-	"help!"				=> \$ep->{run}{help},
-	"colored!"			=> \$ep->{run}{colored},
-	"dummy!"			=> \$ep->{run}{dummy},
-	"verbose!"			=> \$ep->{run}{verbose},
+	"help!"				=> sub { $ep->runner()->help( @_ ); },
+	"colored!"			=> sub { $ep->runner()->colored( @_ ); },
+	"dummy!"			=> sub { $ep->runner()->dummy( @_ ); },
+	"verbose!"			=> sub { $ep->runner()->verbose( @_ ); },
 	"siteSpec!"			=> \$opt_siteSpec,
 	"nodeRoot!"			=> \$opt_nodeRoot,
 	"nodesDirs!"		=> \$opt_nodesDirs,