+++ ChangeLog.md
@@ -9,6 +9,7 @@
     - Reorganize tools/etc vs. site.example/ files (todo #67)
     - Add a comment on how to get site variables (todo #77)
     - ttp.pl purgedirs now makes use of TTP::Path::removeTree(), this one increasing its verbosity
+    - Get rid of (historical) $ep->{run} object, replacing it with $ep->runner() call
 
 ### 4.2.3