+++ ChangeLog.md
@@ -2,7 +2,7 @@
 
 ## ChangeLog
 
-### 4.3.1-rc.0
+### 4.4.0-rc.0
 
     Release date: